% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{createSim.Kernelheaping}
\alias{createSim.Kernelheaping}
\title{Create heaped data for Simulation}
\usage{
createSim.Kernelheaping(n, distribution, rounds, thresholds, offset = 0,
  downbias = 0.5, Beta = 0, ...)
}
\arguments{
\item{n}{sample size}

\item{distribution}{name of the distribution where random sampling is available, e.g. "norm"}

\item{rounds}{rounding values}

\item{thresholds}{rounding thresholds (for Beta=0)}

\item{offset}{certain value added to all observed random samples}

\item{downbias}{bias parameter}

\item{Beta}{acceleration paramter}

\item{...}{additional attributes handed over to "rdistribution" (i.e. rnorm, rgamma,..)}
}
\value{
List of heaped values, true values and input parameters
}
\description{
Create heaped data for Simulation
}

