% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.Kernelheaping}
\alias{plot.Kernelheaping}
\title{Plot Kernel density estimate of heaped data naively and corrected by partly bayesian model}
\usage{
\method{plot}{Kernelheaping}(x, trueX = NULL, ...)
}
\arguments{
\item{x}{Kernelheaping object produced by \code{dheaping} function}

\item{trueX}{optional, if true values X are known (in simulations, for example) the 'Oracle' density estimate is added as well}

\item{...}{additional arguments given to standard plot function}
}
\value{
plot with Kernel density estimates (Naive, Corrected and True (if provided))
}
\description{
Plot Kernel density estimate of heaped data naively and corrected by partly bayesian model
}
