% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{Rprx}
\alias{Rprx}
\title{Conditional Posterior for R_i given X, beta, a, tau}
\usage{
Rprx(rounds, Bias, RR, beta, gridx, postMatrix)
}
\arguments{
\item{rounds}{rounding values}

\item{Bias}{Bias parameter (on probit scale)}

\item{RR}{Threshold values for rounding parameters}

\item{beta}{acceleration parameter}

\item{gridx}{grid on which density is evaluated}

\item{postMatrix}{empty matrix with nrow=2*rounds and ncol=length(gridx) and posterior probabilities to fill}
}
\value{
List with Probabilities
}
\description{
Conditional Posterior for R_i given X, beta, a, tau
}

