% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{Rprx2}
\alias{Rprx2}
\title{Conditional Posterior for X given W, beta, a, p}
\usage{
Rprx2(rounds, Bias, RR, beta, gridx)
}
\arguments{
\item{rounds}{rounding values}

\item{Bias}{Bias parameter (on probit scale)}

\item{RR}{Threshold values for rounding parameters}

\item{beta}{acceleration parameter}

\item{gridx}{grid on which density is evaluated}
}
\value{
List with Probabilities
}
\description{
Conditional Posterior for X given W, beta, a, p
}

