% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simSummary.Kernelheaping}
\alias{simSummary.Kernelheaping}
\title{Simulation Summary}
\usage{
simSummary.Kernelheaping(sim, coverage = 0.9)
}
\arguments{
\item{sim}{Simulation object returned from sim.Kernelheaping}

\item{coverage}{probability for computing coverage intervals}
}
\value{
list with summary statistics
}
\description{
Simulation Summary
}
