% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{students}
\alias{students}
\title{Student0405}
\source{
http://mathfaculty.fullerton.edu/mori/Math120/Data/readme%20datasets/readme.html#Student0405
}
\description{
Data collected during 2004 and 2005 from students in statistics classes 
at a large state university in the northeastern United States.
}
\references{
Utts, J. M., & Heckard, R. F. (2011). Mind on statistics. Cengage Learning.
}
