% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.Kernelheaping}
\alias{summary.Kernelheaping}
\title{Prints some descriptive statistics (means and quantiles) for the estimated rounding, bias and acceleration (beta) parameters}
\usage{
\method{summary}{Kernelheaping}(object, ...)
}
\arguments{
\item{object}{Kernelheaping object produced by \code{dheaping} function}

\item{...}{unused}
}
\value{
Prints summary statistics
}
\description{
Prints some descriptive statistics (means and quantiles) for the estimated rounding, bias and acceleration (beta) parameters
}

