% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tracePlots}
\alias{tracePlots}
\title{Plots some trace plots for the rounding, bias and acceleration (beta) parameters}
\usage{
tracePlots(x, ...)
}
\arguments{
\item{x}{Kernelheaping object produced by \code{dheaping} function}

\item{...}{additional arguments given to standard plot function}
}
\value{
Prints summary statistics
}
\description{
Plots some trace plots for the rounding, bias and acceleration (beta) parameters
}

