% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{students}
\alias{students}
\title{Student0405}
\source{
http://mathfaculty.fullerton.edu/mori/Math120/Data/readme%20datasets/readme.html#Student0405
}
\description{
Data collected during 2004 and 2005 from students in statistics classes
at a large state university in the northeastern United States.
}
\author{
Jessica M. Utts, Robert F. Heckard
}

