% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toOtherShape}
\alias{toOtherShape}
\title{Transfer observations to other shape}
\usage{
toOtherShape(Kest, shapefile)
}
\arguments{
\item{Kest}{Estimation object created by function dshapebivr}

\item{shapefile}{The new shapefile for which the observations shall be transferred to}
}
\value{
The function returns the count of the observations in the different shapfile:
}
\description{
Transfer observations to other shape
}
